gpio-ir: Proof-of-concept tick detector
This adds basic support for a buffered tick detector.

It's based on Marvell's example GPIO IR driver, but measures up-
and down-pulses.  It also uses a timer event rather than a tasklet
to ensure that an entire command is buffered.

It doesn't currently do anything but print events out to the screen.
But for now, that's a good start.