gpio-ir: Refactor IR driver to actually work
The existing IR driver from the vendor had serious issues, including
just plain not working.  This refactors the driver to be more Linuxey.

It is still hardcoded to work with the Nikon remotes we have in-office.

Eventually it would be nice to add Lirc support, but that's a project
for another day.