proc-mohawk: Add fix for L2 cache bug
When executing self-modifying code (such as the Javascript JIT), the L2
cache will sometimes execute the wrong instruction if it's located on a
subsequent page and the page isn't cached.  The solution is to mark address
1023 as non-cacheable.

This code was present in previous releases but was commented out.  This is
because there's a bug where if you use cp14's debug facilities then the L2
cache will never report that it's cleared, and will hang during
initialization.

This patch requires a new defconfig in meta-chumby that disables ICEDCC,
which causes the L2 conflict.