proc-mohawk: Disable cpu_mohawk_proc_fin
The function cpu_mohawk_proc_fin tells the processor to quit,
which involves disabling its caches among other things. Since
we've moved to writethrough, this no longer works.  Because we
don't care about cleaning up when the processor is about to
reboot, this is an acceptable workaround.

In the future, if we decide to try kexec() or something similar,
this will need to be revisited.