pxa168fb-ovly: Add Interlaced mode
This adds an interlaced mode, including a hack to get 1080i working.  It is
untested as to whether or not this hack works on other modes, but
it should work.

It changes the line timings on every frame, adding an extra vblank on
alternate fields.

Note that when programming timings, you must program them for the single-
field case, and set resolution appropriately.  That is, for 1080i mode,
set it so that your field timings are for 540p, but specify a resolution
of 1080 lines.  Also, set vmode to FB_VMODE_INTERLACED.