Revert "pxa168fb-ovly: Enable sign extension"
This reverts commit a405d4ad5028c925f5b492d7e5186b7c67511b8a.

If we enable chroma, we'll have to change the FPGA code in order
to get a new chroma color.  Either that, or modify everything to
use the new chroma color.  Either way, we'll want to modify things
in order to not have a pink screen.

Since we're doing a release, this change will be reverted for now.